app/requirements.txt
fastapi==0.115.8
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
boto3==1.36.17
botocore==1.36.17
uvicorn==0.34.0
SQLAlchemy==2.0.38
asyncmy==0.2.10
magnum==19.0.0